requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
--find-links https://download.pytorch.org/whl/torch_stable.html

annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via starlette
bcrypt==4.2.1
    # via passlib
black==24.10.0
    # via -r requirements.in
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   black
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
contourpy==1.3.0
    # via matplotlib
cryptography==44.0.0
    # via python-jose
cycler==0.12.1
    # via matplotlib
ecdsa==0.19.0
    # via python-jose
fastapi==0.115.5
    # via -r requirements.in
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
flake8==7.1.1
    # via -r requirements.in
fonttools==4.54.1
    # via matplotlib
fsspec==2024.10.0
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via uvicorn
huggingface-hub==0.26.3
    # via timm
idna==3.10
    # via
    #   anyio
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via torch
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.7
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.9.2
    # via
    #   -r requirements.in
    #   seaborn
    #   ultralytics
mccabe==0.7.0
    # via flake8
mpmath==1.3.0
    # via sympy
mypy-extensions==1.0.0
    # via black
networkx==3.4.2
    # via torch
numpy==2.1.3
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
opencv-python==4.10.0.84
    # via
    #   -r requirements.in
    #   ultralytics
packaging==24.1
    # via
    #   black
    #   huggingface-hub
    #   matplotlib
    #   pytest
pandas==2.2.3
    # via
    #   -r requirements.in
    #   seaborn
    #   ultralytics
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pathspec==0.12.1
    # via black
pillow==10.4.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   torchvision
    #   ultralytics
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
psutil==6.1.0
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycodestyle==2.12.1
    # via flake8
pycparser==2.22
    # via cffi
pydantic==2.10.2
    # via fastapi
pydantic-core==2.27.1
    # via pydantic
pyflakes==3.2.0
    # via flake8
pyparsing==3.2.0
    # via matplotlib
pytest==8.3.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.18
    # via -r requirements.in
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   timm
    #   ultralytics
requests==2.32.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   ultralytics
rsa==4.9
    # via python-jose
safetensors==0.4.5
    # via timm
scikit-learn==1.5.2
    # via -r requirements.in
scipy==1.14.1
    # via
    #   scikit-learn
    #   ultralytics
seaborn==0.13.2
    # via ultralytics
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
sympy==1.13.1
    # via torch
threadpoolctl==3.5.0
    # via scikit-learn
timm==1.0.11
    # via -r requirements.in
torch==2.5.1
    # via
    #   -r requirements.in
    #   timm
    #   torchaudio
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchaudio==2.5.1
    # via -r requirements.in
torchvision==0.20.1
    # via
    #   -r requirements.in
    #   timm
    #   ultralytics
tqdm==4.67.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   ultralytics
typing-extensions==4.12.2
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
tzdata==2024.2
    # via pandas
ultralytics==8.3.39
    # via -r requirements.in
ultralytics-thop==2.0.12
    # via ultralytics
urllib3==2.2.3
    # via requests
uvicorn==0.32.1
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
